Actualización clase
El proyecto de conversor no esta terminado, el de classCarpeta si

+++ R2/mvc/src/main/java/com/lmsp/mvc/vista/InterfazVista.java
@@ -17,7 +17,7 @@ public interface InterfazVista {
     void setControlador(ControlConversor c);
     void arranca();
     double getCantidad();
-    int getComision();
+    double getComision();
     void escribeCambio(String s);
     /*
         Constantes para las operaciones
